clarify intent and make static code analysis happy
Add the mandatory class attributes expected of all subclasses to the parent class. This clarifies the intent that they must be set and which type they have by assigning a falsy object of the correct type. A check on instantiation makes sure then, that all values are set now. There are other ways of doing that, but this is the simplest in the current implementation.

Adding an assert isinstance(...) doubles as a sanity check and a help for the IDE to infer the type of an object (therefore append not being flagged as an undefined method on resolved_object_list anymore).