workaround for problem of undefined objects
There is a circular dependency between class objects that are needed during class creation. There is no way to solve this during module import when the class objects are created. An easy way to fix this is to defer the the creation of the data_model dict to a time where all class objects are already created. Her it is deferred to the point in time, when an object of that class is created, simply making data_model an instance object.

As long as not too many objects are created this is the easiest way to ensure that all objects are already created. The fact that there is now one data_model per instance, instead of one per class will only start to be a problem, if this code needs to adhere to strict memory restrictions or millions of objects are created. Then there are ways to make sure that only the first instance that is created, initializes the data_model on the class object and later initializations have no effect. As that is "cleverer" code though, this should only be done if really necessary.